versao recebida da atos em julho

+++ src/pages/admin/profile/patient-documents/index.tsx
@@ -11,10 +11,8 @@ import { useFocusEffect, useNavigation } from '@react-navigation/native'
 import { deleteExam, getPatientExamList } from '@services/exam.service'
 import { CalendarRange, Icon, IconProps, List, ListItem, Modal, Text, useStyleSheet, useTheme } from '@ui-kitten/components'
 import { orderByDateRange, sortByDate } from '@utils/common'
-import React, { FC, ReactElement, RefObject, useCallback, useEffect, useState } from 'react'
-import { Alert, Animated as RNAnimated, ListRenderItemInfo, RefreshControl, TouchableOpacity, View } from 'react-native'
-import { RectButton } from 'react-native-gesture-handler'
-import Swipeable from 'react-native-gesture-handler/Swipeable'
+import React, { FC, ReactElement, useCallback, useEffect, useState } from 'react'
+import { Alert, ListRenderItemInfo, RefreshControl, TouchableOpacity, View } from 'react-native'
 import Animated, { Layout, LightSpeedInLeft, LightSpeedOutRight } from 'react-native-reanimated'
 import Toast from 'react-native-toast-message'
 import { myExamsStyle } from './style'
@@ -71,21 +69,15 @@ const PatientDocumentsScreen: FC = (): ReactElement => {
         }, [])
     )
 
-    const renderRightIcon = (_props: IconProps, exam: ExamDto) => (
-        <View style={styles.viewDate}>
-            <Text
-                style={styles.textDate}
-                appearance='hint'
-                category='c1'>
-                Última atualização
-            </Text>
-            <Text
-                style={styles.textDate}
-                appearance='hint'
-                category='c1'>
-                {localeDateService.format(localeDateService.parse(exam.examResultDate as string, _DATE_FROM_ISO_8601), _DEFAULT_FORMAT_DATE)}
-            </Text>
-        </View>
+    const renderRightIcon = (_props: IconProps, info: ListRenderItemInfo<ExamDto>) => (
+        <>
+            <TouchableOpacity style={{ paddingHorizontal: 15 }} onPress={() => goToDocument(info.item)}>
+                <Icon name='create-outline' pack='ionicons' size={20} />
+            </TouchableOpacity>
+            <TouchableOpacity onPress={() => onDeleteItem(info)}>
+                <Icon name='trash-outline' pack='ionicons' size={20} />
+            </TouchableOpacity>
+        </>
     )
 
     const renderLeftIcon = (props: IconProps) => (
@@ -96,20 +88,33 @@ const PatientDocumentsScreen: FC = (): ReactElement => {
 
     const onDeleteItem = async (info: ListRenderItemInfo<ExamDto>) => {
         try {
-            const resp = await deleteExam(info.item.id)
-            if (resp.status === 201 || resp.status === 200) {
-                const arr = [...data]
-                arr.splice(info.index, 1)
-                if (arr.length === 0)
-                    getExamList()
-                else
-                    setData(arr)
+            Alert.alert(
+                `Deletar "${info.item.examType}"`,
+                `Tem certeza que deseja remover o documento "${info.item.examType}"?`,
+                [
+                    { text: "Não", style: 'cancel', onPress: () => { } },
+                    {
+                        text: 'Sim',
+                        style: 'destructive',
+                        onPress: async () => {
+                            const resp = await deleteExam(info.item.id)
+                            if (resp.status === 201 || resp.status === 200) {
+                                const arr = [...data]
+                                arr.splice(info.index, 1)
+                                if (arr.length === 0)
+                                    getExamList()
+                                else
+                                    setData(arr)
 
-                Toast.show({
-                    type: 'success',
-                    text2: 'Documento deletado',
-                })
-            }
+                                Toast.show({
+                                    type: 'success',
+                                    text2: 'Documento deletado',
+                                })
+                            }
+                        }
+                    },
+                ]
+            )
         } catch (error) {
             Toast.show({
                 type: 'danger',
@@ -130,77 +135,22 @@ const PatientDocumentsScreen: FC = (): ReactElement => {
         getExamList()
     }
 
-    const renderRightAction = (info: ListRenderItemInfo<ExamDto>, text: string,
-        color: string, _progress: RNAnimated.AnimatedInterpolation,
-        dragX: RNAnimated.AnimatedInterpolation, ref: RefObject<Swipeable>) => {
-
-        const opacity = dragX.interpolate({
-            inputRange: [-80, -20, 0],
-            outputRange: [1, 0.9, 0],
-            extrapolate: 'clamp'
-        })
-
-        const pressHandler = () => {
-            ref.current?.close()
-
-            if (text.toUpperCase() === 'DELETAR') {
-                Alert.alert(
-                    `Deletar "${info.item.examType}"`,
-                    `Tem certeza que deseja remover o documento "${info.item.examType}"?`,
-                    [
-                        { text: "Não", style: 'cancel', onPress: () => { } },
-                        {
-                            text: 'Sim',
-                            style: 'destructive',
-                            onPress: () => onDeleteItem(info)
-                        },
-                    ]
-                )
-            } else if (text.toUpperCase() === 'EDITAR') {
-                goToDocument(info.item)
-            }
-        }
-        return (
-            <RNAnimated.View style={[{ flex: 1, transform: [{ translateX: 0 }] }, { opacity: opacity }]}>
-                <RectButton
-                    style={[styles.rightAction, { backgroundColor: theme[color] }]}
-                    onPress={pressHandler}>
-                    <Text style={styles.textWhite}>{text}</Text>
-                </RectButton>
-            </RNAnimated.View>
-        )
-    }
-
     const renderItem = (info: ListRenderItemInfo<ExamDto>) => {
-
-        const ref = React.createRef<Swipeable>()
         return (
-            <Swipeable
-                ref={ref}
-                renderRightActions={(progress, dragX) => (
-                    <View style={styles.viewActions}>
-                        {renderRightAction(info, 'EDITAR', 'color-success-500', progress, dragX, ref)}
-                        {renderRightAction(info, 'DELETAR', 'color-danger-500', progress, dragX, ref)}
-                    </View>
-                )}
-                overshootLeft={false}>
-                <Animated.View
-                    layout={Layout.springify()}
-                    entering={LightSpeedInLeft}
-                    exiting={LightSpeedOutRight}>
-                    <ListItem
-                        style={styles.containerItem}
-                        title={info.item.examType}
-                        description={(evaProps) => (
-                            <>
-                                <Text {...evaProps}>Data: {localeDateService.format(localeDateService.parse(info.item.examDate as string, _DATE_FROM_ISO_8601), _DEFAULT_FORMAT_DATE)}</Text>
-                            </>
-                        )}
-                        accessoryRight={(e) => renderRightIcon(e, info.item)}
-                        accessoryLeft={renderLeftIcon}
-                    />
-                </Animated.View>
-            </Swipeable>
+            <Animated.View
+                layout={Layout.springify()}
+                entering={LightSpeedInLeft}
+                exiting={LightSpeedOutRight}>
+                <ListItem
+                    style={styles.containerItem}
+                    title={info.item.examType}
+                    description={(evaProps) => (
+                        <Text {...evaProps}>Data: {localeDateService.format(localeDateService.parse(info.item.examDate as string, _DATE_FROM_ISO_8601), _DEFAULT_FORMAT_DATE)}</Text>
+                    )}
+                    accessoryRight={(e) => renderRightIcon(e, info)}
+                    accessoryLeft={renderLeftIcon}
+                />
+            </Animated.View>
         )
     }
 
@@ -227,23 +177,25 @@ const PatientDocumentsScreen: FC = (): ReactElement => {
         </View>
     )
 
-    const goToDocument = (data?: ExamDto) => {
-        if (data) {
+    const goToDocument = (exam?: ExamDto) => {
+        if (exam) {
             navigation.navigate('CreatePatientDocuments', {
-                exam: data,
+                exam: exam,
                 props: {
                     editable: true,
                     isNew: false
                 }
             })
         }
-        else navigation.navigate('CreatePatientDocuments', {
-            exam: undefined,
-            props: {
-                editable: true,
-                isNew: true
-            }
-        })
+        else {
+            navigation.navigate('CreatePatientDocuments', {
+                exam: undefined,
+                props: {
+                    editable: true,
+                    isNew: true
+                }
+            })
+        }
     }
 
     return (
